# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API service for Arthemis Brain.",
        "title": "Arthemis Brain API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "http://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
        "/activity/create": {
            "post": {
                "description": "Creates an activity",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "activity"
                ],
                "summary": "Create Activity",
                "parameters": [
                    {
                        "description": "Activity details",
                        "name": "activity",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.ActivityRequest"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Activity created",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    },
                    "400": {
                        "description": "No activities provided",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/activity/delete/{id}": {
            "delete": {
                "description": "Deletes an activity",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "activity"
                ],
                "summary": "Delete Activity",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Activity ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Activity removed successfully",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    "400": {
                        "description": "ID not informed",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Activity not found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/activity/update/{id}": {
            "put": {
                "description": "Updates an activity",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "activity"
                ],
                "summary": "Update Activity",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Activity ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Activity details to update",
                        "name": "project",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ActivityRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Activity updated successfully",
                        "schema": {
                            "$ref": "#/definitions/models.Activity"
                        }
                    },
                    "400": {
                        "description": "ID not informed",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Activity not found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/activity/{id}": {
            "get": {
                "description": "Retrieves an activity",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "activity"
                ],
                "summary": "Get Activity",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Activity ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Activity retrtieved successfully",
                        "schema": {
                            "$ref": "#/definitions/models.Activity"
                        }
                    },
                    "400": {
                        "description": "ID not informed",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Activity not found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
...
        "models.Activity": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "indicators": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Indicator"
                    }
                },
                "justification": {
                    "type": "string"
                },
                "locations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Location"
                    }
                },
                "name": {
                    "type": "string"
                },
                "projectID": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.ActivityRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "justification": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "projectID": {
                    "type": "integer"
                }
            }
        },
        "models.Indicator": {
            "type": "object",
            "properties": {
                "activityID": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "justification": {
                    "type": "string"
                },
                "locationID": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "observationMethod": {
                    "type": "string"
                },
                "observations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Observation"
                    }
                },
                "unit": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "valueBaseline": {
                    "type": "number"
                },
                "valueReference": {
                    "type": "number"
                }
            }
        },
...
        "models.JSONB": {
            "type": "object",
            "additionalProperties": {}
        },
        "models.Location": {
            "type": "object",
            "properties": {
                "country": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "ecosystem": {
                    "type": "string"
                },
                "extent": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "indicators": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Indicator"
                    }
                },
                "position": {
                    "$ref": "#/definitions/models.JSONB"
                },
                "projectID": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "models.Observation": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "indicatorID": {
                    "type": "integer"
                },
                "position": {
                    "$ref": "#/definitions/models.JSONB"
                },
                "updatedAt": {
                    "type": "string"
                },
                "value": {
                    "type": "number"
                }
            }
        },
...
        "utils.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Tarefa não encontrada"
                },
                "message": {
                    "type": "string",
                    "example": "ID inválido fornecido"
                }
            }
        }
    }
}